requirements.txt
pandas==2.0.3
beautifulsoup4~=4.12.3
python-dateutil~=2.8.2
azure-storage-blob==12.19.1
